+++ api/client.go
@@ -288,6 +288,52 @@ func (c *Client) GetTask(p GetTaskParam) (*dto.Task, error) {
 	return task, err
 }
 
+// CreateTimeEntryParam params to create a new time entry
+type CreateTimeEntryParam struct {
+	Workspace   string
+	Start       time.Time
+	End         *time.Time
+	Billable    bool
+	Description string
+	ProjectID   string
+	TaskID      string
+	TagIDs      []string
+}
+
+// CreateTimeEntry create a new time entry
+func (c *Client) CreateTimeEntry(p CreateTimeEntryParam) (dto.TimeEntryImpl, error) {
+	var t dto.TimeEntryImpl
+
+	var end *dto.DateTime
+	if p.End != nil {
+		end = &dto.DateTime{Time: *p.End}
+	}
+
+	r, err := c.NewRequest(
+		"POST",
+		fmt.Sprintf(
+			"workspaces/%s/timeEntries/",
+			p.Workspace,
+		),
+		dto.CreateTimeEntryRequest{
+			Start:       dto.DateTime{Time: p.Start},
+			End:         end,
+			Billable:    p.Billable,
+			Description: p.Description,
+			ProjectID:   p.ProjectID,
+			TaskID:      p.TaskID,
+			TagIDs:      p.TagIDs,
+		},
+	)
+
+	if err != nil {
+		return t, err
+	}
+
+	_, err = c.Do(r, &t)
+
+	return t, err
+}
 
 
 // GetProjectsParam params to get all project of a workspace